1、路况知音项目提交

+++ AdditionFunction/EAIntroView/EAIntroView.m
@@ -265,7 +265,7 @@
 }
 
 - (void)crossDissolveForOffset:(float)offset {
-    NSInteger page = (int)(offset);
+    int page = (int)(offset);
     float alphaValue = offset - (int)offset;
     
     if (alphaValue < 0 && self.currentPageIndex == 0){
